refactor(#1627): extract value_parsing.rs inline tests to sibling file
value_parsing.rs was already over the 800-line source ratchet threshold
(1536 lines); the scalar-decode fix pushed it further. Per the campsite
rule (#1116/#1135) extract the inline `#[cfg(test)] mod tests` body into
value_parsing_tests.rs via `#[path]` (preserving `super::*` semantics),
shrinking value_parsing.rs to 791 lines. No test logic changed.
