+++ cqlite-core/src/storage/sstable/reader/parsing/mod.rs
@@ -9,9 +9,9 @@
 
 // Sub-modules
 mod block_entries;
-mod custom_scalar;
 pub(crate) mod byte_comparable; // Needs to be accessible from row_cell_state_machine
 pub(crate) mod comparator_value_parsing; // Standalone comparator-based parsing for state machine
+mod custom_scalar;
 mod key_parsing;
 mod v5_compressed_legacy;
 mod value_parsing;
